python_packages/pantrysoft/pyproject.toml
[tool.poetry]
name = "pantrysoft"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
selenium = "^4.26.1"
python-dotenv = "^1.0.1"
foodpantry = {path = "../foodpantry", develop = true}


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
